add a unit test
