+++ packages/cli/src/modules/test/commands/repo/test.ts
@@ -422,5 +422,8 @@ export async function command(opts: OptionValues, cmd: Command): Promise<void> {
     };
   }
 
+  console.log('retry');
+  console.log(`args: ${JSON.stringify(args)}`);
+
   await runJest(args);
 }
