+++ packages/core-app-api/report.api.md
@@ -21,6 +21,7 @@ import { AuthRequestOptions } from '@backstage/core-plugin-api';
 import { BackstageIdentityApi } from '@backstage/core-plugin-api';
 import { BackstageIdentityResponse } from '@backstage/core-plugin-api';
 import { BackstagePlugin } from '@backstage/core-plugin-api';
+import { BackstageUserIdentity } from '@backstage/core-plugin-api';
 import { bitbucketAuthApiRef } from '@backstage/core-plugin-api';
 import { bitbucketServerAuthApiRef } from '@backstage/core-plugin-api';
 import { ComponentType } from 'react';
@@ -548,6 +549,9 @@ export class OAuth2
   static create(
     options: OAuth2CreateOptions | OAuth2CreateOptionsWithAuthConnector,
   ): OAuth2;
+  static create(
+    options: OAuth2CreateOptions | OAuth2CreateOptionsWithAuthConnector,
+  ): OAuth2;
   // (undocumented)
   getAccessToken(
     scope?: string | string[],
@@ -563,10 +567,8 @@ export class OAuth2
   getProfile(options?: AuthRequestOptions): Promise<ProfileInfo | undefined>;
   // (undocumented)
   static normalizeScopes(
+    scopeTransform: (scopes: string[]) => string[],
     scopes?: string | string[],
-    options?: {
-      scopeTransform: (scopes: string[]) => string[];
-    },
   ): Set<string>;
   // (undocumented)
   sessionState$(): Observable<SessionState>;
@@ -587,6 +589,22 @@ export type OAuth2CreateOptionsWithAuthConnector = {
   scopeTransform?: (scopes: string[]) => string[];
   defaultScopes?: string[];
   authConnector: AuthConnector<OAuth2Session>;
+};
+
+// @public (undocumented)
+export type OAuth2Response = {
+  providerInfo: {
+    accessToken: string;
+    idToken: string;
+    scope: string;
+    expiresInSeconds?: number;
+  };
+  profile: ProfileInfo;
+  backstageIdentity: {
+    token: string;
+    expiresInSeconds?: number;
+    identity: BackstageUserIdentity;
+  };
 };
 
 // @public
