+++ packages/core-app-api/src/apis/implementations/auth/oauth2/OAuth2.test.ts
@@ -23,6 +23,7 @@ import {
   AuthConnector,
   AuthConnectorRefreshSessionOptions,
   openLoginPopup,
+  // OAuth2Response,
   OAuth2CreateOptionsWithAuthConnector,
 } from '../../../../index';
