+++ packages/core-app-api/src/apis/implementations/auth/oauth2/types.ts
@@ -16,8 +16,11 @@
 
 import {
   BackstageIdentityResponse,
+  BackstageUserIdentity,
   ProfileInfo,
 } from '@backstage/core-plugin-api';
+import { OAuthApiCreateOptions } from '../types.ts';
+import { AuthConnector, PopupOptions } from '../../../../lib';
 import { OAuthApiCreateOptions } from '../types.ts';
 import { AuthConnector, PopupOptions } from '../../../../lib';
 
@@ -36,6 +39,24 @@ export type OAuth2Session = {
   };
   profile: ProfileInfo;
   backstageIdentity?: BackstageIdentityResponse;
+};
+
+/**
+ * @public
+ */
+export type OAuth2Response = {
+  providerInfo: {
+    accessToken: string;
+    idToken: string;
+    scope: string;
+    expiresInSeconds?: number;
+  };
+  profile: ProfileInfo;
+  backstageIdentity: {
+    token: string;
+    expiresInSeconds?: number;
+    identity: BackstageUserIdentity;
+  };
 };
 
 /**
